refactor: extract tailwind arbitrary-value detection into shared lint_utils
