+++ crates/biome_html_analyze/tests/specs/nursery/noTailwindArbitraryValue/invalid.html
@@ -6,3 +6,5 @@
 <div class="text-red-500/[0.31]"></div>
 <div class="[color:red]"></div>
 <div classList="p-[10px]"></div>
+<div CLASS="m-[2px]"></div>
+<div CLASSLIST="p-[10px]"></div>
